Modularize player routes into separate file
