+++ template-engine/index.js
@@ -30,6 +30,7 @@ const {
   updatePlayer,
 } = require('./private/src/teamStorage.js');
 const teamsRoutes = require('./private/src/routing/team.js');
+const playerRoutes = require('./private/src/routing/player.js');
 const { storage, imageFilter } = require('./private/src/multerConfig.js');
 const { ensureLoggedIn, validateUsername } = require('./private/src/auth.js');
 
@@ -86,47 +87,7 @@ app.get('/user/teams', (req, res) => {
     },
   });
 });
-app.route('/user/teams/:teamId/player')
-  .post((req, res) => {
-    const { username } = req.session;
-    const { teamId } = req.params;
-    const { player } = req.body;
-    console.log(`User ${username} added player ${player.name} to team ${teamId}`);
-    try {
-      const newId = addPlayer(username, teamId, player);
-      console.log(`Added player ${newId} to team ${teamId}`);
-      res.status(200).send(newId.toString());
-    } catch (error) {
-      console.log(error);
-      res.status(400).send('Error adding player to team');
-    }
-  })
-  .patch((req, res) => {
-    const { username } = req.session;
-    const { teamId } = req.params;
-    const { player } = req.body;
-    console.log(`User ${username} updated player ${player.id} in team ${teamId}`);
-    try {
-      updatePlayer(username, teamId, player);
-      res.status(204).send();
-    } catch (error) {
-      console.log(error);
-      res.status(400).send('Error updating player in team');
-    }
-  })
-  .delete((req, res) => {
-    const { username } = req.session;
-    const { teamId } = req.params;
-    const { playerId } = req.body;
-    console.log(`User ${username} is removing player ${playerId} from team ${teamId}`);
-    try {
-      removePlayer(username, teamId, playerId);
-      res.status(204).send();
-    } catch {
-      res.status(400).send('Error removing player from team');
-    }
-  });
-app.use('/user/teams', teamsRoutes);
+app.use('/user/teams', playerRoutes);
 app.put('/user/reset/all', (req, res) => {
   const { username } = req.session;
   console.log(`Resetting user ${username}`);
